form works but still need to refine real-time updates on page
